+++ sphinxcontrib/needs/needs.py
@@ -43,7 +43,7 @@ else:
 
     logging.basicConfig()  # Only need to do this once
 
-VERSION = '0.5.4'
+VERSION = '0.5.5'
 
 NEEDS_FUNCTIONS_CONF = []